# Drop sha256-simd in favor of stdlib