@lunny 2023-11-03
But that's only for amd64? I think nowadays, arm64 is also important. Maybe we can wait or have a compile tag.