@wxiaoguang 2023-11-04
1. it doesn't really "drop", the minio package still uses it.
2. according to Go document: `SHA-224 and SHA-256 operations now use native instructions when available when GOARCH=amd64`, the Go package doesn't have arm64 support, while `minio/sha256-simd` has.